% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singscore.R
\name{singsingscoreHCA}
\alias{singsingscoreHCA}
\title{Calculate singscore for specified genes}
\usage{
singsingscoreHCA(
  upSet,
  downSet = NULL,
  tissue = pull(distinct(get_metadata(), tissue_harmonised)),
  cell_type = pull(distinct(get_metadata(), cell_type_harmonised)),
  filter_condition = NULL
)
}
\arguments{
\item{tissue}{A character vector of tissue types to include in the analysis. Default is all tissues in the metadata.}

\item{cell_type}{A character vector of cell types to include in the analysis. Default is all cell types in the metadata.}

\item{filter_condition}{An optional filtering condition to apply to the metadata.}

\item{genes}{A character vector of gene names.}
}
\value{
A data frame with singscore results.
}
\description{
This function calculates the singscore for a given set of genes across specified tissues and cell types.
}
\examples{
singsingscoreHCA(c("CD3G", "CD8A"), tissue = "lung", cell_type = "t_nk")

}
